updated file name changes, and spec changes for rename and deleting data we didnt need for this ticket
