+++ modules/claims_api/app/sidekiq/claims_api/v1/form526_establishment_upload.rb
@@ -1,34 +1,33 @@
 # frozen_string_literal: true
 
+# rubocop:disable Metrics/MethodLength
 require 'pdf_generator_service/pdf_client'
-require 'claims_api/v2/disability_compensation_evss_mapper'
-require 'evss_service/base'
+require 'claims_api/v1/disability_compensation_fes_mapper'
+require 'fes_service/base'
 
 module ClaimsApi
-  module V2
-    class DisabilityCompensationDockerContainerUpload < ClaimsApi::ServiceBase
-      LOG_TAG = '526_v2_Docker_Container_job'
+  module V1
+    class Form526EstablishmentUpload < ClaimsApi::ServiceBase
+      LOG_TAG = '526_v1_establishment_upload'
       sidekiq_options expires_in: 48.hours, retry: true
 
-      def perform(claim_id) # rubocop:disable Metrics/MethodLength
+      def perform(claim_id)
         log_job_progress(claim_id,
                          'Docker container job started')
-
         auto_claim = get_claim(claim_id)
         # Reset for a rerun on this
         set_pending_state_on_claim(auto_claim) unless auto_claim.status == pending_state_value
 
-        evss_data = evss_mapper_service(auto_claim).map_claim
-
         log_job_progress(claim_id,
                          'Submitting mapped data to Docker container')
 
-        evss_res = evss_service.submit(auto_claim, evss_data)
+        fes_data = v1_fes_mapper_service(auto_claim).map_claim
+        fes_res = fes_service.submit(auto_claim, fes_data)
 
         log_job_progress(claim_id,
-                         "Successfully submitted to Docker container with response: #{evss_res}")
+                         "Successfully submitted to Docker container with response: #{fes_res}")
         # update with the evss_id returned
-        auto_claim.update!(evss_id: evss_res[:claimId])
+        auto_claim.update!(evss_id: fes_res[:claimId])
         # clear out the evss_response value on successful submssion to docker container
         clear_evss_response_for_claim(auto_claim)
         # queue flashes job
@@ -49,7 +48,8 @@ module ClaimsApi
                          "Docker container job errored #{e.class}: #{auto_claim&.evss_response}")
         if will_retry?(auto_claim, e)
           raise e
-        else # form526.submit.noRetryError OR form526.InProcess error returned
+        else
+          # form526.submit.noRetryError OR form526.InProcess error returned
           {}
         end
       rescue => e
@@ -61,6 +61,8 @@ module ClaimsApi
 
         raise e
       end
+
+      # rubocop:enable Metrics/MethodLength
 
       private
