+++ modules/claims_api/lib/claims_api/v1/disability_compensation_fes_mapper.rb
@@ -23,32 +23,16 @@ module ClaimsApi
           data: {
             serviceTransactionId: @auto_claim.auth_headers['va_eauth_service_transaction_id'],
             claimantParticipantId: extract_claimant_participant_id,
-            veteranParticipantId: extract_veteran_participant_id,
             form526: @fes_claim
           }
         }
       end
 
-      def claim_attributes
-        # method calls
-      end
+      def claim_attributes; end
 
       def extract_claimant_participant_id
-        # For dependent claims, use dependent participant ID
-        if @auto_claim.auth_headers&.dig('dependent', 'participant_id').present?
-          @auto_claim.auth_headers.dig('dependent', 'participant_id')
-        else
-          # Otherwise, claimant is the veteran
-          extract_veteran_participant_id
-        end
-      end
-
-      def extract_veteran_participant_id
-        # Try auth_headers first, then fall back to other sources
-        # NOTE: veteran_icn is NOT a valid participant ID and would require BGS lookup
-        @auto_claim.auth_headers&.dig('va_eauth_pid') ||
-          @auto_claim.auth_headers&.dig('participant_id') ||
-          @auto_claim.veteran_icn # fallback, would need BGS lookup to convert
+        @auto_claim.auth_headers&.dig('dependent', 'participant_id').present?
+        @auto_claim.auth_headers.dig('dependent', 'participant_id')
       end
     end
   end
